requirements.txt
hydra-core==1.3.2
matplotlib==3.8.3
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
scipy==1.13.0
torch==1.10.1+cu102
torch_geometric==2.5.3
